Change PubSub to an interface
